+++ R/rule_text_bold.R
@@ -74,3 +74,25 @@ cf_field_to_latex.cf_field_rule_text_bold <- function(cf_field, xview, unlocked)
   }
   list(before = before, after = after, unlocked = unlocked)
 }
+
+cf_field_to_gtable.cf_field_rule_text_bold <- function(cf_field, xview, gridobj, unlocked, has_rownames, has_colnames) {
+  css_values <- cf_field[["css_values"]]
+  to_lock <- !is.na(css_values)
+  css_values[is.na(css_values) | !unlocked] <- ""
+  gpbold <- grid::gpar(fontface = "bold")
+
+  row_col <- which(css_values == "bold", arr.ind = TRUE)
+  for (tobold in seq_len(nrow(row_col))) {
+    ind <- find_cell(gridobj,
+                     as.integer(has_colnames) + row_col[tobold, 1],
+                     as.integer(has_rownames) + row_col[tobold, 2],
+                     name = "core-fg")
+    gridobj$grobs[ind][[1]][["gp"]][["fontface"]] <- gpbold[["fontface"]]
+    gridobj$grobs[ind][[1]][["gp"]][["font"]] <- gpbold[["font"]]
+  }
+
+  if (cf_field[["lock_cells"]]) {
+    unlocked <- unlocked | to_lock
+  }
+  list(gridobj = gridobj, unlocked = unlocked)
+}